+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to the Jakarta Migration IntelliJ plugin will be documented 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.13] - 2026-04-12
+
+- Added deep scanning of transitive dependencies
+- Replaced time estimate with confidence score
+
 ## [1.0.12] - 2026-04-09
 
 ### Licensing Changes